+++ core/src/main/java/io/snabble/sdk/Shop.java
@@ -153,6 +153,7 @@ public class Shop implements Serializable, Parcelable {
     private String country;
     private String state;
     private String phone;
+    private String email;
     private boolean isPreLaunch;
     private Map<String, Href> links;
     @SerializedName("lat")
@@ -240,6 +241,13 @@ public class Shop implements Serializable, Parcelable {
     }
 
     /**
+     * Gets the email of the shop
+     */
+    public String getEmail() {
+        return email;
+    }
+
+    /**
      * Gets the latitude of the shop
      */
     public double getLatitude() {
@@ -256,7 +264,7 @@ public class Shop implements Serializable, Parcelable {
     /**
      * Gets the latitude / longitude pair ot the shop as a Location object
      */
-    public Location getLocation() { 
+    public Location getLocation() {
         Location location = new Location("");
         location.setLatitude(latitude);
         location.setLongitude(longitude);