+++ ui/src/main/java/io/snabble/sdk/ui/payment/creditcard/shared/CustomerInfoInputScreen.kt
@@ -1,8 +1,10 @@
 package io.snabble.sdk.ui.payment.creditcard.shared
 
 import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
@@ -11,6 +13,7 @@ import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
@@ -24,7 +27,10 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import io.snabble.sdk.Snabble
 import io.snabble.sdk.ui.R
+import io.snabble.sdk.ui.ThemeManager
 import io.snabble.sdk.ui.cart.shoppingcart.utils.rememberTextFieldManager
 import io.snabble.sdk.ui.payment.creditcard.shared.country.displayName
 import io.snabble.sdk.ui.payment.creditcard.shared.country.domain.models.Address
@@ -44,20 +50,35 @@ internal fun CustomerInfoInputScreen(
     countryItems: List<CountryItem>,
     onBackNavigationClick: () -> Unit,
 ) {
-    var name by remember { mutableStateOf("") }
+    val preFilledData = Snabble.formPrefillData
+    val primaryButtonConfig = ThemeManager.primaryButtonConfig
+    val secondaryButtonConfig = ThemeManager.secondaryButtonConfig
+
+    var name by remember { mutableStateOf(preFilledData?.name.orEmpty()) }
     var intCallingCode by remember { mutableStateOf("") }
     var phoneNumber by remember { mutableStateOf("") }
-    var email by remember { mutableStateOf("") }
-    var street by remember { mutableStateOf("") }
-    var zip by remember { mutableStateOf("") }
-    var city by remember { mutableStateOf("") }
-    var state by remember { mutableStateOf("") }
-    var country: CountryItem by remember { mutableStateOf(countryItems.loadDefaultCountry()) }
+    var email by remember { mutableStateOf(preFilledData?.email.orEmpty()) }
+    var street by remember { mutableStateOf(preFilledData?.street.orEmpty()) }
+    var zip by remember { mutableStateOf(preFilledData?.zip.orEmpty()) }
+    var city by remember { mutableStateOf(preFilledData?.city.orEmpty()) }
+    var state by remember { mutableStateOf(preFilledData?.stateCode.orEmpty()) }
+
+    var country: CountryItem by remember {
+        mutableStateOf(
+            countryItems.loadPreSetCountry(preFilledData?.countryCode)
+                ?: countryItems.loadDefaultCountry()
+        )
+    }
 
     val textFieldManager = rememberTextFieldManager()
 
     val isRequiredStateSet =
-        if (!countryItems.firstOrNull { it.code == country.code }?.stateItems.isNullOrEmpty()) state.isNotEmpty() else true
+        if (!countryItems.firstOrNull { it.code == country.code }?.stateItems.isNullOrEmpty()) {
+            state.isNotEmpty()
+        } else {
+            true
+        }
+
     val areRequiredFieldsSet = listOf(
         name,
         intCallingCode,
@@ -166,7 +187,7 @@ internal fun CustomerInfoInputScreen(
         CountrySelectionMenu(
             countryItems = countryItems,
             selectedCountryCode = country,
-            selectedStateCode = null,
+            selectedStateCode = state,
             onCountrySelected = { countryItem, stateItem ->
                 country = countryItem
                 state = stateItem?.code.orEmpty()
@@ -179,11 +200,16 @@ internal fun CustomerInfoInputScreen(
             verticalArrangement = Arrangement.spacedBy(4.dp),
         ) {
             Button(
-                modifier = Modifier.fillMaxWidth(),
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .defaultMinSize(minHeight = primaryButtonConfig.minHeight.dp),
                 onClick = { onSendAction(createCustomerInfo()) },
                 enabled = !isLoading && areRequiredFieldsSet
             ) {
-                Text(stringResource(R.string.Snabble_Payment_CustomerInfo_next))
+                Text(
+                    stringResource(R.string.Snabble_Payment_CustomerInfo_next),
+                    fontSize = primaryButtonConfig.textSize.sp
+                )
             }
             AnimatedVisibility(visible = showError) {
                 Text(
@@ -194,14 +220,32 @@ internal fun CustomerInfoInputScreen(
                 )
             }
         }
-        TextButton(
-            modifier = Modifier.fillMaxWidth(),
-            onClick = onBackNavigationClick
-        ) {
-            Text(text = stringResource(R.string.Snabble_cancel))
+        if (!secondaryButtonConfig.useOutlinedButton) {
+            TextButton(
+                modifier = Modifier.fillMaxWidth(),
+                onClick = onBackNavigationClick
+            ) {
+                Text(text = stringResource(R.string.Snabble_cancel))
+            }
+        } else {
+            OutlinedButton(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .defaultMinSize(minHeight = secondaryButtonConfig.minHeight.dp),
+                onClick = onBackNavigationClick,
+                border = BorderStroke(width = 1.dp, color = MaterialTheme.colorScheme.primary)
+            ) {
+                Text(
+                    text = stringResource(R.string.Snabble_cancel),
+                    fontSize = secondaryButtonConfig.textSize.sp
+                )
+            }
         }
     }
 }
+
+private fun List<CountryItem>?.loadPreSetCountry(countryCode: String?): CountryItem? =
+    this?.firstOrNull { it.code == countryCode }
 
 private fun List<CountryItem>?.loadDefaultCountry(): CountryItem =
     this?.firstOrNull { it.displayName == Locale.getDefault().country.displayName }