+++ ui/src/main/java/io/snabble/sdk/ui/remotetheme/SnabbleSecondaryButton.kt
@@ -2,8 +2,11 @@ package io.snabble.sdk.ui.remotetheme
 
 import android.content.Context
 import android.content.res.ColorStateList
+import android.graphics.Color
 import android.util.AttributeSet
+import androidx.core.graphics.ColorUtils
 import com.google.android.material.button.MaterialButton
+import com.google.android.material.color.MaterialColors
 import io.snabble.sdk.Project
 import io.snabble.sdk.Snabble
 import io.snabble.sdk.ui.R
@@ -26,28 +29,43 @@ class SnabbleSecondaryButton @JvmOverloads constructor(
 
     private fun setProjectAppTheme() {
         val project = Snabble.checkedInProject.value
+        setRippleColor(project)
         setTextColorFor(project)
+        setStrokeColor(project)
+    }
+
+    private fun setStrokeColor(project: Project?) {
+        strokeColor = ColorStateList.valueOf(context.primaryColorForProject(project))
+    }
+
+    private fun setRippleColor(project: Project?) {
+        val highlightColor = MaterialColors.getColor(
+            context,
+            com.google.android.material.R.attr.colorControlHighlight,
+            Color.TRANSPARENT
+        )
+
+        val rippleColorStateList = ColorStateList.valueOf(
+            ColorUtils.setAlphaComponent(
+                context.primaryColorForProject(project), Color.alpha(highlightColor)
+            )
+        )
+
+        rippleColor = rippleColorStateList
     }
 
     private fun setTextColorFor(project: Project?) {
-        val defaultTextColorStateList = textColors
-
-        // Extract the default disabled color
-        val defaultDisabledTextColor = defaultTextColorStateList.getColorForState(
-                intArrayOf(-android.R.attr.state_enabled),
-                currentTextColor
-            )
-
         val states = arrayOf(
             intArrayOf(-android.R.attr.state_enabled),
-            intArrayOf(android.R.attr.state_enabled)
+            intArrayOf(android.R.attr.state_enabled),
         )
 
         val colors = intArrayOf(
-            defaultDisabledTextColor,
-            context.primaryColorForProject(project)
+            context.primaryColorForProject(project),
+            context.primaryColorForProject(project),
         )
 
         setTextColor(ColorStateList(states, colors))
     }
+
 }